feat(train-simulation): moved train related data to the additionalData field in VehicleData
* updated template for 2023 license header
* adjusted train apps according to new API
