+++ fed/mosaic-sumo/src/main/java/org/eclipse/mosaic/fed/sumo/bridge/traci/constants/CommandRetrieveVehicleState.java
@@ -68,6 +68,7 @@ public class CommandRetrieveVehicleState {
     public final static SumoVar VAR_EMISSIONS_ELECTRICITY = SumoVar.var(0x71);
 
     public final static SumoVar VAR_GET_NEXT_STOPS = SumoVar.var(0x73);
+    public final static SumoVar VAR_GET_LINE = SumoVar.var(0xbd);
 
     public final static SumoVar VAR_STOP_STATE = SumoVar.var(0xb5);
