+++ fed/mosaic-sumo/src/main/java/org/eclipse/mosaic/fed/sumo/bridge/traci/reader/VehicleSubscriptionTraciReader.java
@@ -19,10 +19,9 @@ import org.eclipse.mosaic.fed.sumo.bridge.Bridge;
 import org.eclipse.mosaic.fed.sumo.bridge.api.complex.LeadFollowVehicle;
 import org.eclipse.mosaic.fed.sumo.bridge.api.complex.VehicleSubscriptionResult;
 import org.eclipse.mosaic.fed.sumo.bridge.traci.constants.CommandRetrieveVehicleState;
-import org.eclipse.mosaic.lib.objects.vehicle.StoppingPlace;
+import org.eclipse.mosaic.lib.objects.vehicle.TrainData;
 import org.eclipse.mosaic.lib.util.objects.Position;
 
-import com.google.common.collect.Iterables;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -106,7 +105,9 @@ public class VehicleSubscriptionTraciReader extends AbstractSubscriptionTraciRea
         } else if (varId == CommandRetrieveVehicleState.VAR_MIN_GAP.var) {
             result.minGap = (double) varValue;
         } else if (varId == CommandRetrieveVehicleState.VAR_GET_NEXT_STOPS.var) {
-            result.nextStop = Iterables.getFirst((List<StoppingPlace>) varValue, null);
+            result.nextStops = (List<TrainData.StoppingPlace>) varValue;
+        } else if (varId == CommandRetrieveVehicleState.VAR_GET_LINE.var) {
+            result.line = (String) varValue;
         } else if (varId == CommandRetrieveVehicleState.VAR_LENGTH.var) {
             result.length = (double) varValue;
         } else if (varId == CommandRetrieveVehicleState.VAR_WIDTH.var) {
