+++ legal/templates/license-header-2023.txt
@@ -1,0 +1,12 @@
+Copyright (c) 2023 Fraunhofer FOKUS and others. All rights reserved.
+
+See the NOTICE file(s) distributed with this work for additional
+information regarding copyright ownership.
+
+This program and the accompanying materials are made available under the
+terms of the Eclipse Public License 2.0 which is available at
+http://www.eclipse.org/legal/epl-2.0
+
+SPDX-License-Identifier: EPL-2.0
+
+Contact: mosaic@fokus.fraunhofer.de
